requirements.txt
edn_format
requests
yfinance
pandas-ta
vectorbt
quantstats
pytz